@@ -10,14 +10,14 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^13.3.1",
-    "@angular/common": "^13.3.1",
-    "@angular/compiler": "^13.3.1",
-    "@angular/core": "^13.3.1",
-    "@angular/forms": "^13.3.1",
-    "@angular/platform-browser": "^13.3.1",
-    "@angular/platform-browser-dynamic": "^13.3.1",
-    "@angular/router": "^13.3.1",
+    "@angular/animations": "^14.1.0",
+    "@angular/common": "^14.1.0",
+    "@angular/compiler": "^14.1.0",
+    "@angular/core": "^14.1.0",
+    "@angular/forms": "^14.1.0",
+    "@angular/platform-browser": "^14.1.0",
+    "@angular/platform-browser-dynamic": "^14.1.0",
+    "@angular/router": "^14.1.0",
     "bootstrap": "^5.2.0",
     "jest-editor-support": "*",
     "rxjs": "~7.4.0",
@@ -25,9 +25,9 @@
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^13.3.8",
-    "@angular/cli": "^13.3.1",
-    "@angular/compiler-cli": "^13.3.1",
+    "@angular-devkit/build-angular": "^14.1.0",
+    "@angular/cli": "^14.1.0",
+    "@angular/compiler-cli": "^14.1.0",
     "@types/jasmine": "~3.10.0",
     "@types/node": "^12.11.1",
     "jasmine-core": "~3.10.0",
@@ -36,6 +36,6 @@
     "karma-coverage": "~2.0.3",
     "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "~1.7.0",
-    "typescript": "~4.4.3"
+    "typescript": "~4.7.4"
   }
 }
